[JENKINS-57903] - Breakdown the PromotionProcess class so that the majority of code stays in Promoted Builds
