+++ promotion-api-plugin/src/main/java/hudson/plugins/promoted_builds/Promotion.java
